refactor: move Windows implementations under windows_ops
