+++ src-tauri/src/windows_ops/details/capture/base.rs
