+++ src-tauri/src/windows_ops/details/capture/desktop_dup_window.rs
