+++ src-tauri/src/windows_ops/details/capture/gdi.rs
