+++ src-tauri/src/windows_ops/details/capture/screen_dc.rs
