+++ src-tauri/src/windows_ops/details/mod.rs
@@ -1,0 +1,10 @@
+//! Windows-heavy 实现的私有区域。
+
+pub(super) mod admin;
+pub(super) mod capture;
+pub(super) mod dialog;
+pub(super) mod hotkey;
+mod include;
+pub(super) mod input;
+pub(super) mod webview2;
+pub(super) mod window;
